Move and split RulesetValidator
